Add files via upload
If your website needs proper protection from scammers , it’s recommended to use a CAPTCHA. There are a few extra measure to take when using any CAPTCHA code: 
Secure image , Unique CAPTCHAs , Script Security , Accessibility Therefore , it is beneficial to use CAPTCHA in order to keep your website safe. While the idea of creating your own CAPTCHA sounds ideal, we don’t recommend creating one because of the many failure modes that you may run into. We recommend a website that creates one for you such as reCAPTCHA from Google. In order to keep your device safe in addition to CAPTCHA , be sure to download an antivirus to stay secure.